Fixed bug printing a victory statement for ties
In the playSingleRound function, there was no console message designed specifically for
the case of a tie. Instead, the word "tie" was put into the victory message which didn't make sense.
To fix, I stored the winner in a variable called winner, rather than returning
checkWinner(playerSelection, computerSelction) itself (probably best practice anyway).
Then I used if/else to determine if the round resulted in a tie and the console will now
print the appropriate message in either case.