Added input validation
I added input validation in a while loop inside of the playerSingleRound function.
This will check to make sure that the user inputs a valid option and will prompt the user again
if they have not. One imperfection that remains is that the game still prints out a nonsensical win
statement even if the user entered invalid input. I'm not sure why this would occur before exiting the
do while loop, and the condition to exit the loop is valid input.