Created outer loop that repeats the game as long as the player chooses.
The function is simply called playRockPaperScissors and runs
playSeries in a while loop. At the end of each loop, the user
will be prompted as to whether they would like to play again (y/n).
At this point, the only bug I am finding is in the playSingleRound function
where the console still logs the winner message even if the user's input is
invalid. This does not effect game functionality, but adds clutter to the console.