Remove console-based gameplay and add UI
This has been totally reworked. The playSeries and
playRockPaperScissors functions have been completely disposed
of and the playSingleRound function is now called playRound.
All while loops have been removed, and the functionality
implemented with while loops previously have been added to playRound
via a series of if statements to check if the player or computer
has reached 3 wins. The reason for this restructuring is that
adding the functionality as an event listener was easier without
having to rely on an outside function tracking the wins. The clearest
path forward was to remove those functions altogether.